+++ src/components/ClubDetail/ClubDetailHeader.tsx
@@ -1,15 +1,14 @@
 import styled from 'styled-components';
 import { DEFAULT_IMG } from '@/constants/DEFAULT_IMG';
 import Button from '../Common/Button';
-import {
-    useClubDetail,
-    useClubDetailHeader,
-} from '@/hooks/club-detail/useClubDetail';
+
 import {
     getCategoryEmoji,
     getCategoryMapping,
 } from '@/utils/clubDetail/getCategoryEmoji';
 import { getRecruitmentStatus } from '@/utils/clubDetail/getRecruitmentStatus';
+import { useClubDetail } from '@/hooks/club-detail/useClubDetail';
+import { useClubDetailHeader } from '@/hooks/club-detail/useClubDetailHeader';
 
 interface RecruitStateProps {
     $state?: '모집 중' | '모집 예정' | '모집 마감';
