+++ src/components/ClubDetail/ClubDetailTabs/Recruit/Recruit.tsx
@@ -1,14 +1,21 @@
 import styled from 'styled-components';
 import ContentBlock from '../ContentBlock';
+import { useClubDetail } from '@/hooks/club-detail/useClubDetail';
+import { useClubRecruit } from '@/hooks/club-detail/useClubRecruit';
 
 export default function Recruit() {
     // 훅으로 받은 내용 각 컴포넌트에 props로 전달
+    const { clubId, isPreview } = useClubDetail();
+    const { data } = useClubRecruit(clubId || '', isPreview);
     return (
         <RecruitContainer>
-            <ContentBlock title="모집기간" />
-            <ContentBlock title="모집대상" />
-            <ContentBlock title="유의사항" />
-            <ContentBlock title="💡 기타 동아리 모집 안내" />
+            <ContentBlock title="모집기간" content={data?.due} />
+            <ContentBlock title="모집대상" content={data?.target} />
+            <ContentBlock title="유의사항" content={data?.notice} />
+            <ContentBlock
+                title="💡 기타 동아리 모집 안내"
+                content={data?.etc}
+            />
         </RecruitContainer>
     );
 }
